+++ api/app.py
@@ -1,7 +1,7 @@
 from flask import Flask, render_template, request, url_for, redirect, session
-from app.extra_functions import limit_numeric_to_2_decimals
-from app.data_analysis_classes import DataAnalysis
-from app.data_analysis import remove_initial_lap, preprocess_lap_times
+from api.extra_functions import limit_numeric_to_2_decimals
+#from app.data_analysis_classes import DataAnalysis
+#from app.data_analysis import remove_initial_lap, preprocess_lap_times
 #from app.Read_supabase_data import *
 import pandas as pd
 import os
